Nishi, Vishal #645 Renaming to Data Transfer Objects, updating field types

+++ whp-reports-domain/src/main/java/org/motechproject/whp/reports/domain/patient/Treatment.java
@@ -50,7 +50,7 @@ public class Treatment {
     private String preTreatmentSmearTestResult;
 
     @Column(name = "pretreatment_weight")
-    private int preTreatmentWeight;
+    private Double preTreatmentWeight;
 
     @Column(name = "is_current_treatment")
     private String currentTreatment;
